+++ tests/Pest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 use AdilAzhari\LaravelTrace\Tests\TestCase;
 
 require_once __DIR__.'/Support/readerHelpers.php';
+require_once __DIR__.'/Support/envHelpers.php';
 
 uses(TestCase::class)->in(__DIR__);